Adds map change trigger
Implements a map change trigger that displays a message
when the player collides with a map change object.

This is done by adding collision callbacks to the world
and setting a flag in the player object when a collision
between the player and a map change object occurs.
The draw function of the player then checks this flag
and displays a message if it is set.

Also creates a new world file.